test: 상품 로직 테스트

+++ src/test/java/kitchenpos/test/fixture/ProductFixture.java
@@ -5,10 +5,10 @@ import kitchenpos.domain.Product;
 
 public class ProductFixture {
 
-    public static Product 상품(String name, long price) {
+    public static Product 상품(String name, BigDecimal price) {
         Product product = new Product();
         product.setName(name);
-        product.setPrice(BigDecimal.valueOf(price));
+        product.setPrice(price);
         return product;
     }
 }
